Consolidate interval type reflection into one core helper
The "is this an interval and what are its endpoint types" logic was
duplicated three ways: EfCore's IntervalTypes.IsInterval, the JSON factory's
SupportedDefinitions/StartEndTypes, and OpenApi's TryGetIntervalEndpoints.
Replace all three with StrongTypes.IntervalTypes.IsInterval / TryGetEndpoints
and route the consumers through it; delete the EfCore copy.
